check and cart

+++ client/user/src/api/cartApi.js
@@ -22,22 +22,31 @@ export const getCartByIdApi = async (id) => {
     return res.data;
 }
 
+// created a new cart of user account
 export const addCartApi = async (cart) => {
     const res = await CartApi.post('/api/cart', cart);
     return res.data;
 }
 
-export const updateCartApi = async (cart) => {
-    const res = await CartApi.post(`/api/cart/${cart.id}`, cart);
+// deleted a food item on cart
+export const deleteFoodItemApi = async (foodItem) => {
+    const res = await CartApi.post(`/api/cart/${foodItem.userID}`, foodItem);
     return res.data;
 }
 
+// updated quantity and note
+export const updateCartApi = async (cart) => {
+    const res = await CartApi.post(`/api/cart/uid/${cart.userID}`, cart);
+    return res.data;
+}
+
+// updated or insert (food list of detail)
 export const updateCartFoodApi = async (cart) => {
     const res = await CartApi.post(`/api/cart/food`, cart);
     return res.data;
 }
 
 export const deleteCartApi = async (id) => {
-    const res = await CartApi.delete(`/api/cart/${id}`);
+    const res = await CartApi.delete(`/api/cart/uid/${id}`);
     return res.data;
 }
